+++ src/index.ts
@@ -1,5 +1,5 @@
 export { default as bindContext } from './bind-context';
-export { create, destroy } from './root-lifecycle';
+export { createRoot as create, destroyRoot as destroy } from './root-lifecycle';
 export { createContext } from './inherited-context';
 export type { Resource, ResourceHandle } from './utility-types';
 export type { default as ResourceScope } from './resource-scope';