+++ src/root-lifecycle.ts
@@ -1,9 +1,10 @@
-import { createWithContext } from './resource-lifecycle';
+import { createWithContext, destroy } from './resource-lifecycle';
 import { ParametrizedResourceFactory } from './utility-types';
 import { ResourceFactory } from './utility-types';
+import { resources, roots } from './global-weakrefs';
 
 /** Provision a resource and return its external API. */
-export const create = async <
+export const createRoot = async <
   Controls extends object,
   Args extends Array<unknown>,
 >(
@@ -14,9 +15,22 @@ export const create = async <
   ...args: Args
 ): Promise<Controls> => {
   const rootContext = Object.create(null);
-  return createWithContext(rootContext, factory, ...args);
+  const root = await createWithContext(rootContext, factory, ...args);
+  roots.add(root);
+
+  return root;
 };
 
-// Destroying a root resource is the same process as destroying a child
-// resource. No need to change the implementation.
-export { destroy } from './resource-lifecycle';
+/**
+ * Tear down the resource and all its children, permanently destroying the
+ * reference. This cannot be used to destroy child resources, only roots.
+ */
+export const destroyRoot = async (resource: object): Promise<void> => {
+  if (resources.has(resource) && !roots.has(resource)) {
+    throw new Error(
+      'Cannot destroy child resource. It is owned by another scope.',
+    );
+  }
+
+  return destroy(resource);
+};