+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Renamed public type `ResourceContext` to `ResourceScope`.
+- `destroy(...)` is no longer allowed to destroy child resources, only roots.
 
 ## [0.5.1] - 2023-08-12